Split mobile album actions into focused helper modules
